mine-penge/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
langdetect==1.0.9
python-dotenv==1.0.0
pydantic==2.5.0
lxml-html-clean==0.4.2
aiohttp==3.9.1 